fix bugs ....

+++ src/com/robot/runtime/info/RuntimeInfo.java
@@ -6,7 +6,7 @@ package com.robot.runtime.info;
 //饿汉式单例类.在类初始化时，已经自行实例化
 public class RuntimeInfo {
 
-    private float running_page_count = 0;
+    private int running_page_count = 0;
 
     private int uncommit_word_count = 0;
 
@@ -29,7 +29,7 @@ public class RuntimeInfo {
         System.out.print("running[" + String.format("%6d", running_page_count) + "%] word[" + String.format("%7d", uncommit_word_count) + "] phrase[" + String.format("%7d", uncommit_phrase_count) + "] url[" + String.format("%5d", uncommit_url_count) + "] " + String.format("%s", running_sentence) + "\r");
     }
 
-    public void update_running_page_count(float a_running_page_count) {
+    public void update_running_page_count(int a_running_page_count) {
         running_page_count = a_running_page_count;
         print_runtime_info();
     }
